+++ forui/lib/src/theme/theme_data.dart
@@ -5,7 +5,6 @@ import 'package:meta/meta.dart';
 
 import 'package:forui/forui.dart';
 import 'package:forui/src/theme/breakpoints.dart';
-import 'package:forui/src/widgets/line_calendar/line_calendar.dart';
 
 /// Defines the configuration of the overall visual [FTheme] for a widget subtree.
 ///
@@ -68,9 +67,6 @@ final class FThemeData with Diagnosticable {
 
   /// The label styles.
   final FLabelStyles labelStyles;
-
-  /// The line calendar style.
-  final FLineCalendarStyle lineCalendarStyle;
 
   /// The popover's style.
   final FPopoverStyle popoverStyle;
@@ -136,7 +132,6 @@ final class FThemeData with Diagnosticable {
     required this.dividerStyles,
     required this.headerStyle,
     required this.labelStyles,
-    required this.lineCalendarStyle,
     required this.popoverStyle,
     required this.popoverMenuStyle,
     required this.progressStyle,
@@ -184,7 +179,6 @@ final class FThemeData with Diagnosticable {
       dividerStyles: FDividerStyles.inherit(colorScheme: colorScheme, style: style),
       headerStyle: FHeaderStyles.inherit(colorScheme: colorScheme, typography: typography, style: style),
       labelStyles: FLabelStyles.inherit(style: style),
-      lineCalendarStyle: FLineCalendarStyle.inherit(colorScheme: colorScheme, style: style, typography: typography),
       popoverStyle: FPopoverStyle.inherit(colorScheme: colorScheme, style: style),
       popoverMenuStyle: FPopoverMenuStyle.inherit(colorScheme: colorScheme, style: style, typography: typography),
       progressStyle: FProgressStyle.inherit(colorScheme: colorScheme, style: style),
@@ -262,7 +256,6 @@ final class FThemeData with Diagnosticable {
     FDividerStyles? dividerStyles,
     FHeaderStyles? headerStyle,
     FLabelStyles? labelStyles,
-    FLineCalendarStyle? lineCalendarStyle,
     FPopoverStyle? popoverStyle,
     FPopoverMenuStyle? popoverMenuStyle,
     FProgressStyle? progressStyle,
@@ -295,7 +288,6 @@ final class FThemeData with Diagnosticable {
         dividerStyles: dividerStyles ?? this.dividerStyles,
         headerStyle: headerStyle ?? this.headerStyle,
         labelStyles: labelStyles ?? this.labelStyles,
-        lineCalendarStyle: lineCalendarStyle ?? this.lineCalendarStyle,
         popoverStyle: popoverStyle ?? this.popoverStyle,
         popoverMenuStyle: popoverMenuStyle ?? this.popoverMenuStyle,
         progressStyle: progressStyle ?? this.progressStyle,
@@ -333,7 +325,6 @@ final class FThemeData with Diagnosticable {
       ..add(DiagnosticsProperty('dividerStyles', dividerStyles, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('headerStyle', headerStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('labelStyles', labelStyles, level: DiagnosticLevel.debug))
-      ..add(DiagnosticsProperty('lineCalendarStyle', lineCalendarStyle))
       ..add(DiagnosticsProperty('popoverStyle', popoverStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('popoverMenuStyle', popoverMenuStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('progressStyle', progressStyle, level: DiagnosticLevel.debug))
@@ -372,7 +363,6 @@ final class FThemeData with Diagnosticable {
           dividerStyles == other.dividerStyles &&
           headerStyle == other.headerStyle &&
           labelStyles == other.labelStyles &&
-          lineCalendarStyle == other.lineCalendarStyle &&
           popoverStyle == other.popoverStyle &&
           popoverMenuStyle == other.popoverMenuStyle &&
           progressStyle == other.progressStyle &&
@@ -407,7 +397,6 @@ final class FThemeData with Diagnosticable {
       dividerStyles.hashCode ^
       headerStyle.hashCode ^
       labelStyles.hashCode ^
-      lineCalendarStyle.hashCode ^
       popoverStyle.hashCode ^
       popoverMenuStyle.hashCode ^
       progressStyle.hashCode ^