+++ forui/test/src/widgets/header/root_header_golden_test.dart
@@ -80,6 +80,30 @@ void main() {
 
         await expectLater(find.byType(TestScaffold), matchesGoldenFile('header/root/${theme.name}-focused.png'));
       });
+
+      testWidgets('${theme.name} with RTL FRootHeader actions', (tester) async {
+        await tester.pumpWidget(
+          TestScaffold(
+            theme: theme.data,
+            textDirection: TextDirection.rtl,
+            child: FHeader(
+              title: const Text(title),
+              actions: [
+                FHeaderAction(
+                  icon: FIcon(FAssets.icons.alarmClock),
+                  onPress: null,
+                ),
+                FHeaderAction(
+                  icon: FIcon(FAssets.icons.plus),
+                  onPress: () {},
+                ),
+              ],
+            ),
+          ),
+        );
+
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('header/root/${theme.name}-rtl.png'));
+      });
     }
   });
 }